refactor(friends): refactor screen friends
