--- components/friends/Search-Bar.tsx
@@ -1,46 +1,0 @@
-import React, { useState } from 'react';
-import { View, TextInput, StyleSheet } from 'react-native';
-import Icon from 'react-native-vector-icons/Ionicons';
-
-const SearchBar = ({ onSearch }: { onSearch: (text: string) => void }) => {
-  const [searchText, setSearchText] = useState('');
-
-  const handleTextChange = (text: string) => {
-    setSearchText(text);
-    onSearch(text);
-  };
-
-  return (
-    <View style={styles.container}>
-      <Icon name="search" size={20} color="transparent" style={styles.icon} />
-      <TextInput
-        style={styles.input}
-        placeholder="Rechercher..."
-        value={searchText}
-        onChangeText={handleTextChange}
-      />
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: 'transparent',
-    borderRadius: 8,
-    paddingHorizontal: 10,
-    paddingVertical: 5,
-    margin: 10,
-  },
-  icon: {
-    marginRight: 5,
-  },
-  input: {
-    flex: 1,
-    fontSize: 16,
-    color: 'transparent',
-  },
-});
-
-export default SearchBar;
