+++ components/friends/SearchBar.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import { StyleSheet } from 'react-native';
+import { ThemedView } from '../theme/ThemedView';
+import { ThemedTextInput } from '../theme/ThemedTextInput';
+
+
+export const SearchBar = ({ onSearch }: { onSearch: (text: string) => void }) => (
+    <ThemedView style={styles.searchContainer}>
+      <ThemedTextInput
+        style={styles.searchInput}
+        placeholder="Search for a user..."
+        placeholderTextColor="#aaa"
+        onChangeText={onSearch}
+      />
+    </ThemedView>
+  );
+
+const styles = StyleSheet.create({
+    searchContainer: {
+        padding: 10,
+        backgroundColor: 'transparent', 
+      
+      },
+      searchInput: {
+        backgroundColor: '#222', 
+        color: '#fff', 
+        borderRadius: 8,
+        padding: 10,
+        fontSize: 16,
+      },
+});
+
